Update AutoMaterial_RedShift
Still a work in progress. This MEL script will take any selected geometry in your Maya scene and use the name of each object to generate a Redshift-friendly PBR material, complete with file nodes. Export the resulting object as FBX, load it in Substance, and save all your texture files in the folder specified in the $strDir string.